+++ integration_tests/conftest.py
@@ -46,6 +46,31 @@ async def s3_url():
     return os.environ.get("S3_URL", "http://localhost:9000")
 
 
+@pytest.fixture(scope="session")
+async def postgres_db_url(postgres_url) -> str:
+    from sql_model import test_model
+    from sqlalchemy import create_engine
+    from sqlalchemy import text
+
+    dbname = "cleanpython_test"
+    root_engine = create_engine(
+        f"postgresql+psycopg2://{postgres_url}", isolation_level="AUTOCOMMIT"
+    )
+    with root_engine.connect() as connection:
+        connection.execute(text(f"DROP DATABASE IF EXISTS {dbname}"))
+        connection.execute(text(f"CREATE DATABASE {dbname}"))
+    root_engine.dispose()
+
+    engine = create_engine(
+        f"postgresql+psycopg2://{postgres_url}/{dbname}", isolation_level="AUTOCOMMIT"
+    )
+    with engine.connect() as connection:
+        test_model.metadata.drop_all(engine)
+        test_model.metadata.create_all(engine)
+    engine.dispose()
+    return f"{postgres_url}/{dbname}"
+
+
 def wait_until_url_available(url: str, max_tries=10, interval=0.1):
     # wait for the server to be ready
     for _ in range(max_tries):
