+++ integration_tests/sql_model.py
@@ -1,0 +1,30 @@
+from sqlalchemy import Boolean
+from sqlalchemy import Column
+from sqlalchemy import DateTime
+from sqlalchemy import Float
+from sqlalchemy import Integer
+from sqlalchemy import MetaData
+from sqlalchemy import Table
+from sqlalchemy import Text
+from sqlalchemy import text
+
+test_model = Table(
+    "test_model",
+    MetaData(),
+    Column("id", Integer, primary_key=True, autoincrement=True),
+    Column("t", Text, nullable=False),
+    Column("f", Float, nullable=False),
+    Column("b", Boolean, nullable=False),
+    Column("updated_at", DateTime(timezone=True), nullable=False),
+    Column("n", Float, nullable=True),
+)
+
+
+### For SQLProvider integration tests
+count_query = text("SELECT COUNT(*) FROM test_model")
+insert_query = text(
+    "INSERT INTO test_model (t, f, b, updated_at) "
+    "VALUES ('foo', 1.23, TRUE, '2016-06-22 19:10:25-07') "
+    "RETURNING id"
+)
+update_query = text("UPDATE test_model SET t='bar' WHERE id=:id RETURNING t")
